requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
boto3==1.36.0
botocore==1.36.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
dill==0.3.8
docker-pycreds==0.4.0
faiss-cpu==1.10.0
filelock==3.14.0
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.27.1
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
s3transfer==0.11.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.0
sentry-sdk==2.4.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.3.0
tqdm==4.66.4
transformers==4.48.0
triton==2.3.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
xxhash==3.4.1
yarl==1.9.4
